chore(client-utils): esm
