+++ packages/libs/client-utils/config/api-extractor-coin.json
@@ -3,7 +3,6 @@
  */
 {
   "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-
   /**
    * Optionally specifies another JSON config file that this file extends from.  This provides a way for
    * standard settings to be shared across multiple projects.
@@ -17,7 +16,6 @@
    */
   // "extends": "./shared/api-extractor-base.json"
   // "extends": "my-package/include/api-extractor-base.json"
-
   /**
    * Determines the "<projectFolder>" token that can be used with other config file settings.  The project folder
    * typically contains the tsconfig.json and package.json config files, but the path is user-defined.
@@ -33,7 +31,6 @@
    * DEFAULT VALUE: "<lookup>"
    */
   // "projectFolder": "..",
-
   /**
    * (REQUIRED) Specifies the .d.ts file to be used as the starting point for analysis.  API Extractor
    * analyzes the symbols exported by this module.
@@ -45,8 +42,7 @@
    *
    * SUPPORTED TOKENS: <projectFolder>, <packageName>, <unscopedPackageName>
    */
-  "mainEntryPointFilePath": "<projectFolder>/lib/coin/index.d.ts",
-
+  "mainEntryPointFilePath": "<projectFolder>/lib/esm/coin/index.d.ts",
   /**
    * A list of NPM package names whose exports should be treated as part of this package.
    *
@@ -61,7 +57,6 @@
    * local files for library1.
    */
   "bundledPackages": [],
-
   /**
    * Determines how the TypeScript compiler engine will be invoked by API Extractor.
    */
@@ -101,7 +96,6 @@
      */
     // "skipLibCheck": true,
   },
-
   /**
    * Configures how the API report file (*.api.md) will be generated.
    */
@@ -110,7 +104,6 @@
      * (REQUIRED) Whether to generate an API report.
      */
     "enabled": true,
-
     /**
      * The filename for the API report files.  It will be combined with "reportFolder" or "reportTempFolder" to produce
      * a full file path.
@@ -121,7 +114,6 @@
      * DEFAULT VALUE: "<unscopedPackageName>.api.md"
      */
     "reportFileName": "client-utils-coin.api.md"
-
     /**
      * Specifies the folder where the API report file is written.  The file name portion is determined by
      * the "reportFileName" setting.
@@ -136,7 +128,6 @@
      * DEFAULT VALUE: "<projectFolder>/temp/"
      */
     // "reportFolder": "<projectFolder>/temp/",
-
     /**
      * Specifies the folder where the temporary report file is written.  The file name portion is determined by
      * the "reportFileName" setting.
@@ -152,7 +143,6 @@
      */
     // "reportTempFolder": "<projectFolder>/temp/"
   },
-
   /**
    * Configures how the doc model file (*.api.json) will be generated.
    */
@@ -161,7 +151,6 @@
      * (REQUIRED) Whether to generate a doc model file.
      */
     "enabled": true,
-
     /**
      * The output path for the doc model file.  The file extension should be ".api.json".
      *
@@ -173,7 +162,6 @@
      */
     "apiJsonFilePath": "<projectFolder>/temp/client-utils-coin.api.json"
   },
-
   /**
    * Configures how the .d.ts rollup file will be generated.
    */
@@ -182,7 +170,6 @@
      * (REQUIRED) Whether to generate the .d.ts rollup file.
      */
     "enabled": true,
-
     /**
      * Specifies the output path for a .d.ts rollup file to be generated without any trimming.
      * This file will include all declarations that are exported by the main entry point.
@@ -196,7 +183,6 @@
      * DEFAULT VALUE: "<projectFolder>/dist/<unscopedPackageName>.d.ts"
      */
     "untrimmedFilePath": "<projectFolder>/temp/client-utils-coin.d.ts"
-
     /**
      * Specifies the output path for a .d.ts rollup file to be generated with trimming for an "alpha" release.
      * This file will include only declarations that are marked as "@public", "@beta", or "@alpha".
@@ -208,7 +194,6 @@
      * DEFAULT VALUE: ""
      */
     // "alphaTrimmedFilePath": "<projectFolder>/dist/<unscopedPackageName>-alpha.d.ts",
-
     /**
      * Specifies the output path for a .d.ts rollup file to be generated with trimming for a "beta" release.
      * This file will include only declarations that are marked as "@public" or "@beta".
@@ -220,7 +205,6 @@
      * DEFAULT VALUE: ""
      */
     // "betaTrimmedFilePath": "<projectFolder>/dist/<unscopedPackageName>-beta.d.ts",
-
     /**
      * Specifies the output path for a .d.ts rollup file to be generated with trimming for a "public" release.
      * This file will include only declarations that are marked as "@public".
@@ -234,7 +218,6 @@
      * DEFAULT VALUE: ""
      */
     // "publicTrimmedFilePath": "<projectFolder>/dist/<unscopedPackageName>-public.d.ts",
-
     /**
      * When a declaration is trimmed, by default it will be replaced by a code comment such as
      * "Excluded from this release type: exampleMember".  Set "omitTrimmingComments" to true to remove the
@@ -244,7 +227,6 @@
      */
     // "omitTrimmingComments": true
   },
-
   /**
    * Configures how the tsdoc-metadata.json file will be generated.
    */
@@ -270,7 +252,6 @@
      */
     // "tsdocMetadataFilePath": "<projectFolder>/dist/tsdoc-metadata.json"
   },
-
   /**
    * Specifies what type of newlines API Extractor should use when writing output files.  By default, the output files
    * will be written with Windows-style newlines.  To use POSIX-style newlines, specify "lf" instead.
@@ -279,7 +260,6 @@
    * DEFAULT VALUE: "crlf"
    */
   // "newlineKind": "crlf",
-
   /**
    * Configures how API Extractor reports error and warning messages produced during analysis.
    *
@@ -312,7 +292,6 @@
          * DEFAULT VALUE: "warning"
          */
         "logLevel": "warning"
-
         /**
          * When addToApiReportFile is true:  If API Extractor is configured to write an API report file (.api.md),
          * then the message will be written inside that file; otherwise, the message is instead logged according to
@@ -322,7 +301,6 @@
          */
         // "addToApiReportFile": false
       }
-
       // "TS2551": {
       //   "logLevel": "warning",
       //   "addToApiReportFile": true
@@ -330,7 +308,6 @@
       //
       // . . .
     },
-
     /**
      * Configures handling of messages reported by API Extractor during its analysis.
      *
@@ -343,7 +320,6 @@
         "logLevel": "warning"
         // "addToApiReportFile": false
       }
-
       // "ae-extra-release-tag": {
       //   "logLevel": "warning",
       //   "addToApiReportFile": true
@@ -351,7 +327,6 @@
       //
       // . . .
     },
-
     /**
      * Configures handling of messages reported by the TSDoc parser when analyzing code comments.
      *
@@ -364,7 +339,6 @@
         "logLevel": "warning"
         // "addToApiReportFile": false
       }
-
       // "tsdoc-link-tag-unescaped-text": {
       //   "logLevel": "warning",
       //   "addToApiReportFile": true
