+++ packages/libs/client-utils/package.json
@@ -15,58 +15,55 @@
   ],
   "exports": {
     ".": {
-      "import": "./lib/index.js",
-      "require": "./lib/index.js",
-      "default": "./lib/index.js",
-      "types": "./lib/index.d.ts"
+      "import": "./lib/esm/index.mjs",
+      "require": "./lib/cjs/index.js",
+      "types": "./lib/esm/index.d.ts"
     },
     "./built-in": {
-      "import": "./lib/built-in/index.js",
-      "require": "./lib/built-in/index.js",
-      "default": "./lib/built-in/index.js",
-      "types": "./lib/built-in/index.d.ts"
+      "import": "./lib/esm/built-in/index.mjs",
+      "require": "./lib/cjs/built-in/index.js",
+      "types": "./lib/esm/built-in/index.d.ts"
     },
     "./coin": {
-      "import": "./lib/coin/index.js",
-      "require": "./lib/coin/index.js",
-      "default": "./lib/coin/index.js",
-      "types": "./lib/coin/index.d.ts"
+      "import": "./lib/esm/coin/index.mjs",
+      "require": "./lib/cjs/coin/index.js",
+      "types": "./lib/esm/coin/index.d.ts"
     },
     "./core": {
-      "import": "./lib/core/index.js",
-      "require": "./lib/core/index.js",
-      "default": "./lib/core/index.js",
-      "types": "./lib/core/index.d.ts"
+      "import": "./lib/esm/core/index.mjs",
+      "require": "./lib/cjs/core/index.js",
+      "types": "./lib/esm/core/index.d.ts"
     },
     "./nodejs": {
-      "import": "./lib/nodejs/index.js",
-      "require": "./lib/nodejs/index.js",
-      "default": "./lib/nodejs/index.js",
-      "types": "./lib/nodejs/index.d.ts"
+      "import": "./lib/esm/nodejs/index.mjs",
+      "require": "./lib/cjs/nodejs/index.js",
+      "types": "./lib/esm/nodejs/index.d.ts"
     },
     "./marmalade": {
-      "import": "./lib/marmalade/index.js",
-      "require": "./lib/marmalade/index.js",
-      "default": "./lib/marmalade/index.js",
-      "types": "./lib/marmalade/index.d.ts"
+      "import": "./lib/esm/marmalade/index.mjs",
+      "require": "./lib/cjs/marmalade/index.js",
+      "types": "./lib/esm/marmalade/index.d.ts"
     }
   },
   "typesVersions": {
     "*": {
+      ".": [
+        "./lib/esm/index.d.ts"
+      ],
       "built-in": [
-        "./lib/built-in/index.d.ts"
+        "./lib/esm/built-in/index.d.ts"
       ],
       "core": [
-        "./lib/core/index.d.ts"
+        "./lib/esm/core/index.d.ts"
       ],
       "coin": [
-        "./lib/coin/index.d.ts"
+        "./lib/esm/coin/index.d.ts"
       ],
       "nodejs": [
-        "./lib/nodejs/index.d.ts"
+        "./lib/esm/nodejs/index.d.ts"
       ],
       "marmalade": [
-        "./lib/marmalade/index.d.ts"
+        "./lib/esm/marmalade/index.d.ts"
       ]
     }
   },
@@ -83,7 +80,7 @@
     "ae:built-in": "api-extractor run --verbose -c ./config/api-extractor-built-in.json",
     "ae:coin": "api-extractor run --verbose -c ./config/api-extractor-coin.json",
     "ae:core": "api-extractor run --verbose -c ./config/api-extractor-core.json",
-    "build": "pnpm run pactjs:generate:contract && pnpm run generate-pipe-type && tsc && pnpm run ae",
+    "build": "pnpm run pactjs:generate:contract && pnpm run generate-pipe-type && unbuild && pnpm run ae",
     "dev:ae": "pnpm run dev:ae:core && pnpm run dev:ae:built-in && pnpm run dev:ae:coin",
     "dev:ae:built-in": "api-extractor run --local --verbose -c ./config/api-extractor-built-in.json",
     "dev:ae:coin": "api-extractor run --local --verbose -c ./config/api-extractor-coin.json",
@@ -131,6 +128,7 @@
     "prettier": "~3.0.3",
     "ts-node": "~10.8.2",
     "typescript": "5.2.2",
+    "unbuild": "^2.0.0",
     "vitest": "^1.1.0"
   }
 }
