+++ packages/libs/toolbox/build.config.ts
@@ -1,24 +1,4 @@
-import { readdirSync } from 'fs';
-import { BuildEntry, defineBuildConfig } from 'unbuild';
-
-const entries: BuildEntry[] = readdirSync('./src').flatMap((entry) => {
-  return [
-    {
-      builder: 'mkdist',
-      input: `./src/${entry}`,
-      outDir: `./dist/${entry}/esm`,
-      cleanDist: true,
-      format: 'esm',
-    },
-    {
-      builder: 'mkdist',
-      input: `./src/${entry}`,
-      outDir: `./dist/${entry}/cjs`,
-      cleanDist: true,
-      format: 'cjs',
-    },
-  ];
-});
+import { defineBuildConfig } from 'unbuild';
 
 export default defineBuildConfig({
   entries: [
